Update Java version and exclude specific files in pom.xml
The Java version requirement in pom.xml was adjusted from 19 to 17 to accommodate projects that are running on a lower Java version. Also, some file paths were added to the exclusions in the checkstyle and PMD configurations to ignore generated sources in test cases and to handle certain code quality checks more efficiently.